# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Simple Inventory API",
    "description": "This is a simple API",
    "contact": {
      "email": "you@your-company.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/VITYLJATROJAN/nika-gold/2.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
...
  "paths": {
...
    "/api/admin/finance/online/orders/{id}/sale": {
      "put": {
        "summary": "Конвертація онлайн-замовлення у продаж",
        "description": "Перетворює завершене замовлення (`status: completed`) у продаж та оновлює фінансові показники.",
        "tags": ["Finance - Online"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-замовлення для конвертації",
            "schema": {
              "type": "string",
              "example": "664a20b4f9a39c51a2d124ab"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успішна конвертація замовлення у продаж",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Sale processed successfully!"
                    },
                    "sale": {
                      "$ref": "#/components/schemas/OnlineSale"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Замовлення повинне бути завершене перед конвертацією",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Order must be completed before converting to sale"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Замовлення не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Order not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера під час обробки продажу",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to process sale"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/finance/online/sales": {
      "get": {
        "tags": ["Finance - Online"],
        "summary": "Отримати всі онлайн-продажі",
        "description": "Повертає список усіх онлайн-продажів із інформацією про продукти, кількість, метод оплати та оброблювача продажу.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список онлайн-продажів успішно отримано.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "description": "ID продажу" },
                      "products": {
                        "type": "array",
                        "description": "Список продуктів у продажу.",
                        "items": {
                          "type": "object",
                          "properties": {
                            "productId": {
                              "type": "string",
                              "description": "ID продукту"
                            },
                            "quantity": {
                              "type": "integer",
                              "description": "Кількість"
                            },
                            "salePrice": {
                              "type": "number",
                              "description": "Ціна продажу за одиницю"
                            }
                          }
                        }
                      },
                      "totalAmount": {
                        "type": "number",
                        "description": "Загальна сума продажу"
                      },
                      "paymentMethod": {
                        "type": "string",
                        "description": "Метод оплати"
                      },
                      "processedBy": {
                        "type": "string",
                        "description": "ID адміністратора, який обробив продаж"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Токен відсутній або недійсний." },
          "403": {
            "description": "Доступ заборонено для неадміністративних користувачів."
          }
        }
      },
      "post": {
        "tags": ["Finance - Online"],
        "summary": "Створити новий онлайн-продаж",
        "description": "Додає новий запис про онлайн-продаж із вказаними продуктами, кількістю, методами оплати і оброблювачем.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "array",
                    "description": "Продукти у продажу.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість"
                        },
                        "salePrice": {
                          "type": "number",
                          "description": "Ціна продажу за одиницю"
                        }
                      }
                    }
                  },
                  "totalAmount": {
                    "type": "number",
                    "description": "Загальна сума продажу"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Метод оплати"
                  },
                  "processedBy": {
                    "type": "string",
                    "description": "ID адміністратора"
                  }
                },
                "example": {
                  "products": [
                    { "productId": "12345", "quantity": 2, "salePrice": 250 },
                    { "productId": "67890", "quantity": 1, "salePrice": 300 }
                  ],
                  "totalAmount": 800,
                  "paymentMethod": "card",
                  "processedBy": "admin123"
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Онлайн-продаж успішно створено." },
          "400": { "description": "Недійсні дані у запиті." }
        }
      }
    },
    "/api/admin/finance/online/sales/{id}": {
      "patch": {
        "tags": ["Finance - Online"],
        "summary": "Оновити дані онлайн-продажу",
        "description": "Оновлює вказані дані конкретного онлайн-продажу за ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-продажу.",
            "schema": { "type": "string", "example": "sale123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "totalAmount": {
                    "type": "number",
                    "description": "Оновлена загальна сума"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Оновлений метод оплати"
                  },
                  "products": {
                    "type": "array",
                    "description": "Оновлений список продуктів.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість"
                        },
                        "salePrice": {
                          "type": "number",
                          "description": "Ціна продажу за одиницю"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "totalAmount": 900,
                  "paymentMethod": "bank_transfer",
                  "products": [
                    { "productId": "12345", "quantity": 1, "salePrice": 250 },
                    { "productId": "67890", "quantity": 2, "salePrice": 300 }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Дані онлайн-продажу успішно оновлено." },
          "400": { "description": "Недійсні дані у запиті." },
          "404": { "description": "Онлайн-продаж не знайдено." }
        }
      }
    },
    "/api/admin/finance/online/sales/{id}/return": {
      "put": {
        "tags": ["Finance - Online"],
        "summary": "Повернення онлайн-продажу",
        "description": "Оновлює статус продажу на `returned`, повертає товари на склад і коригує фінансові показники.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID онлайн-продажу, що повертається.",
            "schema": { "type": "string", "example": "sale123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refundAmount": {
                    "type": "number",
                    "description": "Сума повернення покупцю"
                  }
                },
                "example": { "refundAmount": 500 }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Онлайн-продаж успішно повернуто." },
          "400": { "description": "Недійсні дані у запиті." },
          "404": { "description": "Онлайн-продаж не знайдено." }
        }
      }
    },
...
    "/api/admin/finance/offline/sales": {
      "get": {
        "tags": ["Finance - Offline"],
        "summary": "Отримати всі офлайн-продажі",
        "description": "Повертає список усіх офлайн-продажів із інформацією про продукти, кількість, методи оплати та оброблювачів.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список офлайн-продажів успішно отримано.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "description": "ID продажу" },
                      "products": {
                        "type": "array",
                        "description": "Список продуктів у продажу.",
                        "items": {
                          "type": "object",
                          "properties": {
                            "productId": {
                              "type": "string",
                              "description": "ID продукту"
                            },
                            "quantity": {
                              "type": "integer",
                              "description": "Кількість"
                            },
                            "price": {
                              "type": "number",
                              "description": "Ціна продажу за одиницю"
                            }
                          }
                        }
                      },
                      "totalAmount": {
                        "type": "number",
                        "description": "Загальна сума продажу"
                      },
                      "paymentMethod": {
                        "type": "string",
                        "description": "Метод оплати"
                      },
                      "saleLocation": {
                        "type": "string",
                        "description": "Місце продажу"
                      },
                      "processedBy": {
                        "type": "string",
                        "description": "ID адміністратора, який обробив продаж"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Токен відсутній або недійсний." },
          "403": {
            "description": "Доступ заборонено для неадміністративних користувачів."
          }
        }
      },
      "post": {
        "tags": ["Finance - Offline"],
        "summary": "Створити новий офлайн-продаж",
        "description": "Додає новий запис про офлайн-продаж із вказаними продуктами, кількістю, методами оплати і місцем продажу.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "products": {
                    "type": "array",
                    "description": "Продукти у продажу.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість продукту"
                        },
                        "price": {
                          "type": "number",
                          "description": "Ціна продажу за одиницю"
                        }
                      }
                    }
                  },
                  "totalAmount": {
                    "type": "number",
                    "description": "Загальна сума продажу"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Метод оплати"
                  },
                  "saleLocation": {
                    "type": "string",
                    "description": "Місце продажу"
                  },
                  "processedBy": {
                    "type": "string",
                    "description": "ID адміністратора"
                  }
                },
                "example": {
                  "products": [
                    { "productId": "12345", "quantity": 2, "price": 250 },
                    { "productId": "67890", "quantity": 1, "price": 300 }
                  ],
                  "totalAmount": 800,
                  "paymentMethod": "cash",
                  "saleLocation": "Shop #45",
                  "processedBy": "admin123"
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Офлайн-продаж успішно створено." },
          "400": { "description": "Недійсні дані у запиті." }
        }
      }
    },
    "/api/admin/finance/offline/sales/{id}": {
      "patch": {
        "tags": ["Finance - Offline"],
        "summary": "Оновити дані офлайн-продажу",
        "description": "Оновлює вказані дані конкретного офлайн-продажу за ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID офлайн-продажу.",
            "schema": { "type": "string", "example": "sale123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "totalAmount": {
                    "type": "number",
                    "description": "Оновлена загальна сума"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "description": "Оновлений метод оплати"
                  },
                  "products": {
                    "type": "array",
                    "description": "Оновлений список продуктів.",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "ID продукту"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Кількість"
                        },
                        "price": {
                          "type": "number",
                          "description": "Ціна продажу за одиницю"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "totalAmount": 900,
                  "paymentMethod": "bank_transfer",
                  "products": [
                    { "productId": "12345", "quantity": 1, "price": 250 },
                    { "productId": "67890", "quantity": 2, "price": 300 }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Дані офлайн-продажу успішно оновлено." },
          "400": { "description": "Недійсні дані у запиті." },
          "404": { "description": "Офлайн-продаж не знайдено." }
        }
      }
    },
    "/api/admin/finance/offline/sales/{id}/return": {
      "put": {
        "tags": ["Finance - Offline"],
        "summary": "Повернення офлайн-продажу",
        "description": "Оновлює статус продажу на `returned`, повертає товари на склад і коригує фінансові показники.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID офлайн-продажу, що повертається.",
            "schema": { "type": "string", "example": "sale123" }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "description": "Bearer-токен для авторизації адміністратора.",
            "schema": {
              "type": "string",
              "example": "Bearer eyJhbGciOiJI..."
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refundAmount": {
                    "type": "number",
                    "description": "Сума повернення покупцю"
                  }
                },
                "example": { "refundAmount": 500 }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Офлайн-продаж успішно повернуто." },
          "400": { "description": "Недійсні дані у запиті." },
          "404": { "description": "Офлайн-продаж не знайдено." }
        }
      }
    },
...
  },
...
}